Move filters to main module; apply deadtime to eventlist objects; test thoroughly
